Prepared for merge into master.
